started moving from float64 to float32 for rasterx

+++ rasterx/bezier_test.go
@@ -31,7 +31,7 @@ func CubeLerpTo(ax, ay, bx, by, cx, cy, dx, dy float32, LineTo func(ex, ey float
 	}
 	if devsq >= 0.333 {
 		const tol = 3
-		n := 1 + int(math.Sqrt(math.Sqrt(tol*float64(devsq))))
+		n := 1 + int(math.Sqrt(math.Sqrt(tol*float32(devsq))))
 		t, nInv := float32(0), 1/float32(n)
 		for i := 0; i < n-1; i++ {
 			t += nInv
@@ -51,7 +51,7 @@ func QuadLerpTo(ax, ay, bx, by, cx, cy float32, LineTo func(dx, dy float32)) {
 	devsq := devSquared(ax, ay, bx, by, cx, cy)
 	if devsq >= 0.333 {
 		const tol = 3
-		n := 1 + int(math.Sqrt(math.Sqrt(tol*float64(devsq))))
+		n := 1 + int(math.Sqrt(math.Sqrt(tol*float32(devsq))))
 		t, nInv := float32(0), 1/float32(n)
 		for i := 0; i < n-1; i++ {
 			t += nInv
@@ -153,12 +153,12 @@ func TestBezierCube(t *testing.T) {
 		}
 		//t.Log("Bez to", len(r1x), "lines")
 		for i, v := range r1x {
-			if math.Abs(float64(v-r2x[i])) > epsilon {
+			if math.Abs(float32(v-r2x[i])) > epsilon {
 				t.Error("x mismatch", v, "vs", r2x[i], " diff ", v-r2x[i])
 			}
 		}
 		for i, v := range r1y {
-			if math.Abs(float64(v-r2y[i])) > epsilon {
+			if math.Abs(float32(v-r2y[i])) > epsilon {
 				t.Error("y mismatch", v, "vs", r2y[i], " diff ", v-r2y[i])
 			}
 		}
@@ -194,12 +194,12 @@ func TestBezierQuad(t *testing.T) {
 		}
 		//t.Log("Bez to", len(r1x), "lines")
 		for i, v := range r1x {
-			if math.Abs(float64(v-r2x[i])) > epsilon {
+			if math.Abs(float32(v-r2x[i])) > epsilon {
 				t.Error("x mismatch", v, "vs", r2x[i], " diff ", v-r2x[i])
 			}
 		}
 		for i, v := range r1y {
-			if math.Abs(float64(v-r2y[i])) > epsilon {
+			if math.Abs(float32(v-r2y[i])) > epsilon {
 				t.Error("y mismatch", v, "vs", r2y[i], " diff ", v-r2y[i])
 			}
 		}
